Intermediate step: add desktop target and refactor a lot of stuff to commonMain.
