+++ build.gradle.kts
@@ -1,16 +1,11 @@
 plugins {
-    alias(libs.plugins.android.application).apply(false)
-    alias(libs.plugins.android.library).apply(false)
-    alias(libs.plugins.kotlinx.serialization).apply(false)
-    alias(libs.plugins.kotlin.multiplatform).apply(false)
-    alias(libs.plugins.kotlin.android).apply(false)
-    alias(libs.plugins.kotlin.parcelize).apply(false)
-    alias(libs.plugins.dependencyUpdates).apply(false)
-    alias(libs.plugins.compose.compiler).apply(false)
-}
-
-allprojects {
-    // ./gradlew dependencyUpdates
-    // Report: build/dependencyUpdates/report.txt
-    apply(plugin = "com.github.ben-manes.versions")
+    // this is necessary to avoid the plugins to be loaded multiple times
+    // in each subproject's classloader
+    alias(libs.plugins.android.application) apply false
+    alias(libs.plugins.android.library) apply false
+    alias(libs.plugins.compose.multiplatform) apply false
+    alias(libs.plugins.compose.compiler) apply false
+    alias(libs.plugins.kotlin.parcelize) apply false
+    alias(libs.plugins.kotlin.multiplatform) apply false
+    alias(libs.plugins.kotlinx.serialization) apply false
 }
