+++ composeApp/src/androidMain/kotlin/com/github/jetbrains/rssreader/App.kt
@@ -1,9 +1,9 @@
 package com.github.jetbrains.rssreader
 
 import android.app.Application
+import android.content.Context
 import com.github.jetbrains.rssreader.app.FeedStore
-import com.github.jetbrains.rssreader.core.RssReader
-import com.github.jetbrains.rssreader.core.create
+import com.github.jetbrains.rssreader.core.createAndroid
 import com.github.jetbrains.rssreader.sync.RefreshWorker
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
@@ -20,7 +20,7 @@ class App : Application() {
     }
 
     private val appModule = module {
-        single { RssReader.create(get(), BuildConfig.DEBUG) }
+        single { createAndroid(get<Context>(), BuildConfig.DEBUG) }
         single { FeedStore(get()) }
     }
