+++ composeApp/src/commonMain/kotlin/com/github/jetbrains/rssreader/Settings.kt
@@ -1,0 +1,5 @@
+package com.github.jetbrains.rssreader
+
+class Settings(val defaultFeedUrls: Set<String>) {
+    fun isDefault(feedUrl: String) = defaultFeedUrls.contains(feedUrl)
+}
