+++ composeApp/src/commonMain/kotlin/com/github/jetbrains/rssreader/core/HttpClient.kt
@@ -1,0 +1,23 @@
+package com.github.jetbrains.rssreader.core
+
+import io.github.aakira.napier.Napier
+import io.ktor.client.HttpClient
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.client.plugins.logging.LogLevel
+import io.ktor.client.plugins.logging.Logger
+import io.ktor.client.plugins.logging.Logging
+import io.ktor.serialization.kotlinx.xml.xml
+
+fun HttpClient(withLog: Boolean) = HttpClient() {
+    if (withLog) install(Logging) {
+        level = LogLevel.HEADERS
+        logger = object : Logger {
+            override fun log(message: String) {
+                Napier.v(tag = "HttpClient", message = message)
+            }
+        }
+    }
+    install(ContentNegotiation) {
+        xml()
+    }
+}
