+++ composeApp/src/commonMain/kotlin/com/github/jetbrains/rssreader/ui/PostList.kt
@@ -1,0 +1,99 @@
+package com.github.jetbrains.rssreader.ui
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.LazyListState
+import androidx.compose.foundation.lazy.itemsIndexed
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardElevation
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import coil3.compose.rememberAsyncImagePainter
+import com.github.jetbrains.rssreader.entity.Post
+import kotlinx.datetime.LocalDateTime
+import kotlinx.datetime.TimeZone
+import kotlinx.datetime.TimeZone.Companion.currentSystemDefault
+import kotlinx.datetime.UtcOffset
+import kotlinx.datetime.format
+import kotlinx.datetime.toLocalDateTime
+import kotlin.time.ExperimentalTime
+import kotlinx.datetime.Instant
+import kotlinx.datetime.format.*
+
+@Composable
+fun PostList(
+    modifier: Modifier,
+    posts: List<Post>,
+    listState: LazyListState,
+    onClick: (Post) -> Unit
+) {
+    LazyColumn(
+        modifier = modifier,
+        contentPadding = PaddingValues(16.dp),
+        state = listState
+    ) {
+        itemsIndexed(posts) { i, post ->
+            if (i == 0) Spacer(Modifier.windowInsetsTopHeight(WindowInsets.statusBars))
+            PostItem(post) { onClick(post) }
+            if (i != posts.size - 1) Spacer(modifier = Modifier.size(16.dp))
+        }
+    }
+}
+
+@OptIn(ExperimentalTime::class)
+@Composable
+fun PostItem(
+    item: Post,
+    onClick: () -> Unit
+) {
+    val padding = 16.dp
+    Box {
+        Card( //todo check elevation
+            shape = RoundedCornerShape(padding)
+        ) {
+            Column(
+                modifier = Modifier.clickable(onClick = onClick)
+            ) {
+                Spacer(modifier = Modifier.size(padding))
+                Text(
+                    modifier = Modifier.padding(start = padding, end = padding),
+                    style = MaterialTheme.typography.headlineSmall,
+                    text = item.title
+                )
+                item.imageUrl?.let { url ->
+                    Spacer(modifier = Modifier.size(padding))
+                    Image(
+                        painter = rememberAsyncImagePainter(url),
+                        modifier = Modifier.height(180.dp).fillMaxWidth(),
+                        contentDescription = null
+                    )
+                }
+                item.desc?.let { desc ->
+                    Spacer(modifier = Modifier.size(padding))
+                    Text(
+                        modifier = Modifier.padding(start = padding, end = padding),
+                        style = MaterialTheme.typography.bodyMedium,
+                        maxLines = 5,
+                        overflow = TextOverflow.Ellipsis,
+                        text = desc
+                    )
+                }
+                Spacer(modifier = Modifier.size(padding))
+                Text(
+                    modifier = Modifier.padding(start = padding, end = padding),
+                    style = MaterialTheme.typography.bodySmall,
+                    color = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.7f),
+                    text = Instant.fromEpochMilliseconds(item.date).toString()
+                )
+                Spacer(modifier = Modifier.size(padding))
+            }
+        }
+    }
+}
