+++ settings.gradle.kts
@@ -8,6 +8,12 @@ pluginManagement {
         gradlePluginPortal()
         mavenCentral()
     }
+    plugins {
+        kotlin("jvm") version "2.0.21"
+    }
+}
+plugins {
+    id("org.gradle.toolchains.foojay-resolver-convention") version "0.8.0"
 }
 
 dependencyResolutionManagement {
@@ -17,6 +23,4 @@ dependencyResolutionManagement {
     }
 }
 
-
-include(":shared")
 include(":composeApp")
