--- shared/build.gradle.kts
@@ -1,69 +1,0 @@
-import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
-import org.jetbrains.kotlin.gradle.dsl.JvmTarget
-
-plugins {
-    alias(libs.plugins.android.library)
-    alias(libs.plugins.kotlinx.serialization)
-    alias(libs.plugins.kotlin.multiplatform)
-}
-
-kotlin {
-    androidTarget {
-        @OptIn(ExperimentalKotlinGradlePluginApi::class)
-        compilerOptions {
-            jvmTarget.set(JvmTarget.JVM_11)
-        }
-    }
-
-    listOf(
-        iosX64(),
-        iosArm64(),
-        iosSimulatorArm64()
-    ).forEach {
-        it.binaries.framework {
-            baseName = "RssReader"
-            isStatic = true
-        }
-    }
-
-    sourceSets {
-        commonMain.dependencies {
-            //Network
-            implementation(libs.ktor.core)
-            implementation(libs.ktor.logging)
-            //Coroutines
-            implementation(libs.kotlinx.coroutines.core)
-            //Logger
-            implementation(libs.napier)
-            //JSON
-            implementation(libs.kotlinx.serialization.json)
-            //Key-Value storage
-            implementation(libs.multiplatform.settings)
-            // DI
-            api(libs.koin.core)
-        }
-
-        androidMain.dependencies {
-            //Network
-            implementation(libs.ktor.client.okhttp)
-        }
-
-        iosMain.dependencies {
-            //Network
-            implementation(libs.ktor.client.ios)
-        }
-    }
-}
-
-android {
-    namespace = "org.example.project.shared"
-    compileSdk = libs.versions.android.compileSdk.get().toInt()
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_11
-        targetCompatibility = JavaVersion.VERSION_11
-    }
-    defaultConfig {
-        minSdk = libs.versions.android.minSdk.get().toInt()
-    }
-}
-
