--- shared/src/androidMain/kotlin/com/github/jetbrains/rssreader/core/AndroidHttpClient.kt
@@ -1,24 +1,0 @@
-package com.github.jetbrains.rssreader.core
-
-import io.github.aakira.napier.Napier
-import io.ktor.client.*
-import io.ktor.client.engine.okhttp.*
-import io.ktor.client.plugins.logging.*
-import java.util.concurrent.TimeUnit
-
-internal fun AndroidHttpClient(withLog: Boolean) = HttpClient(OkHttp) {
-    engine {
-        config {
-            retryOnConnectionFailure(true)
-            connectTimeout(5, TimeUnit.SECONDS)
-        }
-    }
-    if (withLog) install(Logging) {
-        level = LogLevel.HEADERS
-        logger = object : Logger {
-            override fun log(message: String) {
-                Napier.v(tag = "AndroidHttpClient", message = message)
-            }
-        }
-    }
-}
