--- shared/src/commonMain/kotlin/com/github/jetbrains/rssreader/core/Settings.kt
@@ -1,5 +1,0 @@
-package com.github.jetbrains.rssreader.core
-
-class Settings(val defaultFeedUrls: Set<String>) {
-    fun isDefault(feedUrl: String) = defaultFeedUrls.contains(feedUrl)
-}
