--- shared/src/commonMain/kotlin/com/github/jetbrains/rssreader/core/datasource/network/FeedLoader.kt
@@ -1,16 +1,0 @@
-package com.github.jetbrains.rssreader.core.datasource.network
-
-import com.github.jetbrains.rssreader.core.entity.Feed
-import io.ktor.client.*
-import io.ktor.client.request.*
-import io.ktor.client.statement.*
-
-internal class FeedLoader(
-    private val httpClient: HttpClient,
-    private val parser: FeedParser
-) {
-    suspend fun getFeed(url: String, isDefault: Boolean): Feed {
-        val xml = httpClient.get(url).bodyAsText()
-        return parser.parse(url, xml, isDefault)
-    }
-}
